average is a floating number

+++ .ccls-cache/@home@runner@CSCI-HW5/HW5-2@main2.cpp
@@ -37,8 +37,8 @@ while (inFile >> dataTemp)
   }
     cout << "The sum of valid data is: " << sumData << endl;
     cout << "The number of valid data points is: " << counter << endl;
-    cout << "The average of the valid data is : " << sumData / counter << endl;
-  
+    cout << "The average of the valid data is : " << (static_cast<double>(sumData) / counter) << endl;
+
   
   
   }
